@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
+        <TargetFramework>net6.0</TargetFramework>
 
         <IsPackable>true</IsPackable>
         <GenerateDocumentationFile>true</GenerateDocumentationFile>
@@ -13,7 +13,7 @@
             Repo: https://github.com/nunohpinheiro/toolpack-exceptions
         </Description>
 
-        <PackageVersion>1.0.0</PackageVersion>
+        <PackageVersion>2.0.0</PackageVersion>
 
         <PackageId>ToolPack.Exceptions.Base</PackageId>
         <PackageProjectUrl>https://github.com/nunohpinheiro/toolpack-exceptions</PackageProjectUrl>
@@ -31,13 +31,7 @@
     </ItemGroup>
 
     <ItemGroup>
-        <PackageReference Include="FluentValidation" Version="10.2.3" />
-        <PackageReference Include="Grpc.AspNetCore.Server" Version="2.38.0" />
-        <PackageReference Include="Grpc.Core.Api" Version="2.38.1" />
-        <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.2.0" />
-        <PackageReference Include="Microsoft.AspNetCore.Mvc.Core" Version="2.2.5" />
-        <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="5.0.0" />
-        <PackageReference Include="System.Threading.Tasks" Version="4.3.0" />
+        <PackageReference Include="FluentValidation" Version="11.0.2" />
     </ItemGroup>
 
 </Project>
