@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
+        <TargetFramework>net6.0</TargetFramework>
 
         <IsPackable>true</IsPackable>
         <GenerateDocumentationFile>true</GenerateDocumentationFile>
@@ -14,7 +14,7 @@
             Repo: https://github.com/nunohpinheiro/toolpack-exceptions
         </Description>
 
-        <PackageVersion>1.0.0</PackageVersion>
+        <PackageVersion>2.0.0</PackageVersion>
 
         <PackageId>ToolPack.Exceptions.Web</PackageId>
         <PackageProjectUrl>https://github.com/nunohpinheiro/toolpack-exceptions</PackageProjectUrl>
@@ -32,13 +32,9 @@
     </ItemGroup>
 
     <ItemGroup>
-        <PackageReference Include="FluentValidation" Version="10.2.3" />
-        <PackageReference Include="Grpc.AspNetCore.Server" Version="2.38.0" />
-        <PackageReference Include="Grpc.Core.Api" Version="2.38.1" />
-        <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.2.0" />
-        <PackageReference Include="Microsoft.AspNetCore.Mvc.Core" Version="2.2.5" />
-        <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="5.0.0" />
-        <PackageReference Include="System.Threading.Tasks" Version="4.3.0" />
+        <PackageReference Include="Grpc.AspNetCore.Server" Version="2.46.0" />
+        <PackageReference Include="Grpc.Core.Api" Version="2.46.3" />
+        <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="7.0.0-preview.4.22229.4" />
     </ItemGroup>
 
     <ItemGroup>
