@@ -1,15 +1,14 @@
-namespace RestAPI
+namespace RestAPI;
+
+using System;
+
+public class WeatherForecast
 {
-    using System;
+    public DateTime Date { get; set; }
 
-    public class WeatherForecast
-    {
-        public DateTime Date { get; set; }
+    public int TemperatureC { get; set; }
 
-        public int TemperatureC { get; set; }
+    public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
 
-        public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
-
-        public string Summary { get; set; }
-    }
+    public string Summary { get; set; }
 }
