@@ -1,103 +1,102 @@
-namespace ToolPack.Exceptions.Web.Models
+namespace ToolPack.Exceptions.Web.Models;
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using ToolPack.Exceptions.Base.Entities;
+
+/// <summary>
+/// Problem Details entity that extends the one in AspNetCore.Mvc (https://docs.microsoft.com/en-us/dotnet/api/microsoft.aspnetcore.mvc.problemdetails).
+/// Used to return application problems, using the RFC 7807 (https://tools.ietf.org/html/rfc7807).
+/// </summary>
+public class ProblemDetails : Microsoft.AspNetCore.Mvc.ProblemDetails
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using ToolPack.Exceptions.Base.Entities;
+    /// <summary>Gets or sets additional errors related with the Problem Details.</summary>
+    public IDictionary<string, string[]> Errors { get; private set; }
 
-    /// <summary>
-    /// Problem Details entity that extends the one in AspNetCore.Mvc (https://docs.microsoft.com/en-us/dotnet/api/microsoft.aspnetcore.mvc.problemdetails).
-    /// Used to return application problems, using the RFC 7807 (https://tools.ietf.org/html/rfc7807).
-    /// </summary>
-    public class ProblemDetails : Microsoft.AspNetCore.Mvc.ProblemDetails
+    /// <summary>Tracing identifier related with the context of these ProblemDetails.</summary>
+    public string TraceId { get; init; }
+
+    /// <summary>Creates a ProblemDetails instance with properties derived from a given exception.</summary>
+    /// <param name="exception">Exception from which the ProblemDetails properties are filled.</param>
+    public ProblemDetails(Exception exception)
     {
-        /// <summary>Gets or sets additional errors related with the Problem Details.</summary>
-        public IDictionary<string, string[]> Errors { get; private set; }
+        var exceptionErrorStatus = WebErrorStatuses.GetFromException(exception);
 
-        /// <summary>Tracing identifier related with the context of these ProblemDetails.</summary>
-        public string TraceId { get; init; }
+        Detail = exception?.Message;
+        Instance = exceptionErrorStatus.Description;
+        Status = (int)exceptionErrorStatus.HttpCode;
+        Title = exceptionErrorStatus.HttpCode.ToString();
+        Type = exceptionErrorStatus.Description;
 
-        /// <summary>Creates a ProblemDetails instance with properties derived from a given exception.</summary>
-        /// <param name="exception">Exception from which the ProblemDetails properties are filled.</param>
-        public ProblemDetails(Exception exception)
+        SetErrors(exception);
+    }
+
+    /// <summary>Creates a ProblemDetails instance with properties derived from a given exception.</summary>
+    /// <param name="exception">Exception from which the ProblemDetails properties are filled.</param>
+    /// <param name="traceId">Tracing identifier related with the context of this ProblemDetails.</param>
+    public ProblemDetails(Exception exception, string traceId)
+        : this(exception)
+    {
+        TraceId = traceId;
+    }
+
+    /// <summary>Creates a default ProblemDetails instance.</summary>
+    /// <param name="traceId">Tracing identifier related with the context of these ProblemDetails.</param>
+    public ProblemDetails(string traceId)
+    {
+        var problemDetailsStatus = WebErrorStatuses.InternalUnknownError;
+
+        Detail = problemDetailsStatus.HttpCode.ToString();
+        Status = (int)problemDetailsStatus.HttpCode;
+        Title = problemDetailsStatus.HttpCode.ToString();
+        TraceId = traceId;
+        Type = problemDetailsStatus.Description;
+    }
+
+    private void SetErrors(Exception exception)
+    {
+        if (exception is null)
+            return;
+
+        if (exception is AggregateException aggregateException)
+            SetErrorsFromAggregateException(aggregateException);
+
+        else if ((exception is ValidationFailedException validationException) && validationException.Errors.Any())
+            Errors = validationException.Errors;
+
+        else if (exception.InnerException is not null)
+            SetErrorsFromInnerException(exception.InnerException);
+    }
+
+    private void SetErrorsFromAggregateException(AggregateException exception)
+    {
+        var failureGroups = exception?.InnerExceptions?.GroupBy(
+                                e => string.IsNullOrWhiteSpace(e.Source) ? "UnknownSource" : e.Source,
+                                e => e.Message);
+
+        if (failureGroups?.Any() is not true)
+            return;
+
+        Errors = new Dictionary<string, string[]>();
+
+        foreach (var failureGroup in failureGroups)
         {
-            var exceptionErrorStatus = WebErrorStatuses.GetFromException(exception);
+            var propertyName = failureGroup.Key;
+            var propertyFailures = failureGroup.ToArray();
 
-            Detail = exception?.Message;
-            Instance = exceptionErrorStatus.Description;
-            Status = (int)exceptionErrorStatus.HttpCode;
-            Title = exceptionErrorStatus.HttpCode.ToString();
-            Type = exceptionErrorStatus.Description;
-
-            SetErrors(exception);
+            Errors.Add(propertyName, propertyFailures);
         }
+    }
 
-        /// <summary>Creates a ProblemDetails instance with properties derived from a given exception.</summary>
-        /// <param name="exception">Exception from which the ProblemDetails properties are filled.</param>
-        /// <param name="traceId">Tracing identifier related with the context of this ProblemDetails.</param>
-        public ProblemDetails(Exception exception, string traceId)
-            : this(exception)
+    private void SetErrorsFromInnerException(Exception innerException)
+    {
+        Errors = new Dictionary<string, string[]>
         {
-            TraceId = traceId;
-        }
-
-        /// <summary>Creates a default ProblemDetails instance.</summary>
-        /// <param name="traceId">Tracing identifier related with the context of these ProblemDetails.</param>
-        public ProblemDetails(string traceId)
-        {
-            var problemDetailsStatus = WebErrorStatuses.InternalUnknownError;
-
-            Detail = problemDetailsStatus.HttpCode.ToString();
-            Status = (int)problemDetailsStatus.HttpCode;
-            Title = problemDetailsStatus.HttpCode.ToString();
-            TraceId = traceId;
-            Type = problemDetailsStatus.Description;
-        }
-
-        private void SetErrors(Exception exception)
-        {
-            if (exception is null)
-                return;
-
-            if (exception is AggregateException aggregateException)
-                SetErrorsFromAggregateException(aggregateException);
-
-            else if ((exception is ValidationFailedException validationException) && validationException.Errors.Any())
-                Errors = validationException.Errors;
-
-            else if (exception.InnerException is not null)
-                SetErrorsFromInnerException(exception.InnerException);
-        }
-
-        private void SetErrorsFromAggregateException(AggregateException exception)
-        {
-            var failureGroups = exception?.InnerExceptions?.GroupBy(
-                                    e => string.IsNullOrWhiteSpace(e.Source) ? "UnknownSource" : e.Source,
-                                    e => e.Message);
-
-            if (failureGroups?.Any() is not true)
-                return;
-
-            Errors = new Dictionary<string, string[]>();
-
-            foreach (var failureGroup in failureGroups)
             {
-                var propertyName = failureGroup.Key;
-                var propertyFailures = failureGroup.ToArray();
-
-                Errors.Add(propertyName, propertyFailures);
+                string.IsNullOrWhiteSpace(innerException.Source) ? "UnknownSource" : innerException.Source,
+                new string[1] { innerException.Message }
             }
-        }
-
-        private void SetErrorsFromInnerException(Exception innerException)
-        {
-            Errors = new Dictionary<string, string[]>
-            {
-                {
-                    string.IsNullOrWhiteSpace(innerException.Source) ? "UnknownSource" : innerException.Source,
-                    new string[1] { innerException.Message }
-                }
-            };
-        }
+        };
     }
 }
