@@ -1,32 +1,31 @@
-namespace ToolPack.Exceptions.UnitTests.Web.Extensions
+namespace ToolPack.Exceptions.UnitTests.Web.Extensions;
+
+using FluentAssertions;
+using NUnit.Framework;
+using System;
+using ToolPack.Exceptions.Web.Extensions;
+
+public class JsonExtensionsTests
 {
-    using FluentAssertions;
-    using NUnit.Framework;
-    using System;
-    using ToolPack.Exceptions.Web.Extensions;
-
-    public class JsonExtensionsTests
+    [Test]
+    public void TrySerializeCamelCase_InputSerializable_ReturnsTrue_OutputStringMatches_OutputExceptionIsNotNull()
     {
-        [Test]
-        public void TrySerializeCamelCase_InputSerializable_ReturnsTrue_OutputStringMatches_OutputExceptionIsNotNull()
-        {
-            // Arrange
-            SomeClass input = new() { SomeProperty = 1 };
+        // Arrange
+        SomeClass input = new() { SomeProperty = 1 };
 
-            string expectedJson = "{\"someProperty\":1}";
+        string expectedJson = "{\"someProperty\":1}";
 
-            // Act
-            var result = input.TrySerializeCamelCase(out string outputString, out Exception outputException);
+        // Act
+        var result = input.TrySerializeCamelCase(out string outputString, out Exception outputException);
 
-            // Assert
-            result.Should().BeTrue();
-            outputString.Should().Be(expectedJson);
-            outputException.Should().BeNull();
-        }
+        // Assert
+        result.Should().BeTrue();
+        outputString.Should().Be(expectedJson);
+        outputException.Should().BeNull();
+    }
 
-        private class SomeClass
-        {
-            public int SomeProperty { get; set; }
-        }
+    private class SomeClass
+    {
+        public int SomeProperty { get; set; }
     }
 }
