@@ -1,63 +1,57 @@
-namespace ToolPack.Exceptions.UnitTests.Web.Handlers
+namespace ToolPack.Exceptions.UnitTests.Web.Handlers;
+
+using FluentAssertions;
+using Microsoft.AspNetCore.Http;
+using Microsoft.Extensions.Logging;
+using Moq;
+using NUnit.Framework;
+using System;
+using System.Threading.Tasks;
+using ToolPack.Exceptions.Web.Handlers;
+using ToolPack.Exceptions.Web.Services.Interfaces;
+
+public class GlobalExceptionMiddlewareTests
 {
-    using FluentAssertions;
-    using Microsoft.AspNetCore.Http;
-    using Microsoft.Extensions.Logging;
-    using Moq;
-    using NUnit.Framework;
-    using System;
-    using System.Threading.Tasks;
-    using ToolPack.Exceptions.Web.Handlers;
-    using ToolPack.Exceptions.Web.Models;
-    using ToolPack.Exceptions.Web.Services.Interfaces;
+    private static Mock<HttpContext> HttpContextMock => new();
+    private static Mock<ILogger<GlobalExceptionMiddleware>> LoggerMock => new();
+    private static Mock<IProblemDetailsService> ProblemDetailsSvcMock => new();
 
-    public class GlobalExceptionMiddlewareTests
+    [Test]
+    public void InvokeAsync_RequestDelegateNotThrowsException_ProblemDetailsServiceIsNotCalled()
     {
-        private static Mock<HttpContext> HttpContextMock => new();
-        private static Mock<ILogger<GlobalExceptionMiddleware>> LoggerMock => new();
-        private static Mock<IProblemDetailsService> ProblemDetailsSvcMock => new();
+        // Arrange
+        var requestDelegate = new RequestDelegate(innerContext => Task.FromResult(0));
+        var exceptionsMiddleware = new GlobalExceptionMiddleware(LoggerMock.Object, requestDelegate);
 
-        [Test]
-        public void InvokeAsync_RequestDelegateNotThrowsException_ProblemDetailsServiceIsNotCalled()
-        {
-            // Arrange
-            var requestDelegate = new RequestDelegate(innerContext => Task.FromResult(0));
-            var exceptionsMiddleware = new GlobalExceptionMiddleware(LoggerMock.Object, requestDelegate);
+        var problemDetailsSvcMock = ProblemDetailsSvcMock;
 
-            var problemDetailsSvcMock = ProblemDetailsSvcMock;
+        // Act
+        Func<Task> act = async () => await exceptionsMiddleware.InvokeAsync(HttpContextMock.Object, problemDetailsSvcMock.Object);
 
-            WebErrorStatus errorStatus = It.IsAny<WebErrorStatus>();
+        // Assert
+        act.Should().NotThrowAsync();
+        problemDetailsSvcMock.Verify(
+            x => x.BuildProblemDetailsResponse(It.IsAny<Exception>()),
+            Times.Never);
+    }
 
-            // Act
-            Func<Task> act = async () => await exceptionsMiddleware.InvokeAsync(HttpContextMock.Object, problemDetailsSvcMock.Object);
+    [Test]
+    public void InvokeAsync_RequestDelegateThrowsException_ProblemDetailsServiceIsCalled()
+    {
+        // Arrange
+        var exception = new Exception("sample message");
+        var requestDelegate = new RequestDelegate(innerContext => throw exception);
+        var exceptionsMiddleware = new GlobalExceptionMiddleware(LoggerMock.Object, requestDelegate);
 
-            // Assert
-            act.Should().NotThrow();
-            problemDetailsSvcMock.Verify(
-                x => x.BuildProblemDetailsResponse(It.IsAny<Exception>()),
-                Times.Never);
-        }
+        var problemDetailsSvcMock = ProblemDetailsSvcMock;
 
-        [Test]
-        public void InvokeAsync_RequestDelegateThrowsException_ProblemDetailsServiceIsCalled()
-        {
-            // Arrange
-            var exception = new Exception("sample message");
-            var requestDelegate = new RequestDelegate(innerContext => throw exception);
-            var exceptionsMiddleware = new GlobalExceptionMiddleware(LoggerMock.Object, requestDelegate);
+        // Act
+        Func<Task> act = async () => await exceptionsMiddleware.InvokeAsync(HttpContextMock.Object, problemDetailsSvcMock.Object);
 
-            var problemDetailsSvcMock = ProblemDetailsSvcMock;
-
-            WebErrorStatus errorStatus = It.IsAny<WebErrorStatus>();
-
-            // Act
-            Func<Task> act = async () => await exceptionsMiddleware.InvokeAsync(HttpContextMock.Object, problemDetailsSvcMock.Object);
-
-            // Assert
-            act.Should().NotThrow();
-            problemDetailsSvcMock.Verify(
-                x => x.BuildProblemDetailsResponse(It.IsAny<Exception>()),
-                Times.Once);
-        }
+        // Assert
+        act.Should().NotThrowAsync();
+        problemDetailsSvcMock.Verify(
+            x => x.BuildProblemDetailsResponse(It.IsAny<Exception>()),
+            Times.Once);
     }
 }
